+++ IceFishing/Controllers/PeopleSearchViewController.swift
@@ -61,7 +61,7 @@ class PeopleSearchViewController: UITableViewController, UITableViewDelegate, UI
     }
     
     override func tableView(tableView: UITableView, heightForRowAtIndexPath indexPath: NSIndexPath) -> CGFloat {
-        return CGFloat(80)
+        return 80
     }
     
     override func tableView(tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
