+++ IceFishing/Controllers/PostHistoryTableViewController.swift
@@ -74,7 +74,7 @@ class PostHistoryTableViewController: UITableViewController, UIScrollViewDelegat
     }
     
     override func tableView(tableView: UITableView, heightForRowAtIndexPath indexPath: NSIndexPath) -> CGFloat {
-        return CGFloat(80)
+        return 80
     }
     
     override func tableView(tableView: UITableView, didSelectRowAtIndexPath indexPath: NSIndexPath) {
