+++ IceFishing/Profile/FollowingViewController.swift
@@ -62,7 +62,7 @@ class FollowingViewController: UITableViewController, UIScrollViewDelegate {
     }
 
     override func tableView(tableView: UITableView, heightForRowAtIndexPath indexPath: NSIndexPath) -> CGFloat {
-        return CGFloat(80)
+        return 80
     }
     
     override func tableView(tableView: UITableView, didSelectRowAtIndexPath indexPath: NSIndexPath) {
